Use Node IP Address instead of Node.Name in minion.ResourceLocation. Refactor GetNodeHostIP into pkg/util/node (instead of pkg/util to break import cycle).
Include internalIP in gce NodeAddresses.
